# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Active Home Platform API
  version: 1.0.0
  summary: Configuration-driven housing & construction advertising marketplace (Uzbekistan) — v1 contract.
  description: "Enterprise OpenAPI 3.1 contract for **Active Home v1**, the implementation contract between\
    \ the\nweb frontend, backend (FastAPI modular monolith), AI coding agents, QA automation, future mobile\n\
    apps, and external integrations.\n\nThis document is generated from and fully aligned with the immutable\
    \ approved architecture:\nSRS v1.0, DDD Domain Model v1.0, Configuration & Metadata Framework v1.0,\
    \ SAD v1.0,\nDatabase Architecture & Logical Data Model v1.0, Physical Database Design v1.0, and the\n\
    API Architecture & Integration Specification v1.0.\n\n## Cross-cutting contract rules (see the API\
    \ Architecture Specification)\n- **Versioning:** URI major versioning under `/api/v1`. Within a major\
    \ version only additive,\n  backward-compatible change is allowed. Configuration changes (new categories,\
    \ forms, fields,\n  plans, roles, templates) are **never** contract changes — dynamic data flows through\
    \ stable\n  envelopes (`AttributeMap`, `LocalizedText`, `FormDefinition`), so the endpoint surface\
    \ is\n  invariant under the platform's normal mode of change.\n- **Authentication:** server-side sessions\
    \ backed by Redis (SAD §13). A session is established by an\n  auth flow and carried by a secure,\
...
  contact:
...
  license:
...
servers:
...
security:
- sessionCookie: []
...
paths:
...
  /search:
    get:
      tags:
      - Search
      operationId: searchListingsGet
      summary: Search listings (query parameters)
      description: 'Full-text, faceted, geo/radius discovery over visible listings. Free text is matched
        cross-script

        (Latin↔Cyrillic, FR-SRCH-004). Facets are the facet-eligible fields selected in the

        SearchConfiguration snapshot. Promoted results are labelled and capped per page (BRULE-10, I-17).

        **Degradation:** if the index is unavailable, results are served from the DB fallback with

        `degraded:true` (QR-05). **Rate limiting:** public anti-scraping limits apply (NFR-SEC-005).

        **Performance:** target < 500 ms p95 (NFR-PERF-001).

        '
      security: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: categoryId
        in: query
        schema:
          type: string
          format: uuid
      - name: listingType
        in: query
        schema:
          type: string
          enum:
          - ADVERTISEMENT
          - PRODUCT
          - SERVICE
      - name: filters
        in: query
        description: Facet filters as repeated `filters[fieldCode]=value` pairs (facet-eligible fields
          only).
        style: deepObject
        explode: true
        schema:
          type: object
          additionalProperties: true
      - name: priceMin
        in: query
        schema:
          type: string
      - name: priceMax
        in: query
        schema:
          type: string
      - name: verifiedOnly
        in: query
        schema:
          type: boolean
      - name: lat
        in: query
        schema:
          type: number
          format: double
      - name: lng
        in: query
        schema:
          type: number
          format: double
      - name: radiusKm
        in: query
        schema:
          type: number
          minimum: 0.1
          maximum: 200
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - RELEVANCE
          - RECENCY
          - PRICE_ASC
          - PRICE_DESC
          default: RELEVANCE
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results with facets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
              examples:
                results:
                  $ref: '#/components/examples/SearchResultExample'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceDegraded'
    post:
      tags:
      - Search
      operationId: searchListingsPost
      summary: Search listings (structured body)
      description: Structured equivalent of GET /search for complex filter payloads. Same semantics and
        degradation behaviour.
      security: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results with facets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceDegraded'
  /search/facets:
    get:
      tags:
      - Search
      operationId: getFacets
      summary: Get available facets for a scope
      description: Returns the facet dimensions (with localized labels) available for a category scope,
        from the SearchConfiguration snapshot.
      security: []
      parameters:
      - name: categoryId
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Facet metadata.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facet'
  /search/suggest:
    get:
      tags:
      - Search
      operationId: suggest
      summary: Autocomplete / suggestions
      description: Query completions and suggestions from the normalized index (cross-script). For type-ahead
        UIs.
      security: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
      responses:
        '200':
          description: Suggestions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Suggestion'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ah_session
      description: 'First-party, server-side session (Redis-backed, SAD §13). Issued by any auth flow
        and sent as a

        secure, HttpOnly, SameSite=Lax cookie. This is the default scheme for browser clients.

        '
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'The same first-party session token presented as `Authorization: Bearer <token>` for
        native and

        mobile clients that cannot use cookies. The token is an **opaque server-side session reference**,

        not a JWT. Revoking the session (logout) invalidates it immediately.

        '
    googleOAuth:
      type: oauth2
      description: 'Google is a federated **sign-in** identity provider only (authorization-code flow).
        After Google

        authenticates the user, Active Home links or creates the account and issues its own first-party

        session (`sessionCookie`/`sessionBearer`). Active Home does not expose an OAuth2 authorization

        server of its own.

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect sign-in
...
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque forward cursor from a previous page's `nextCursor`. Omit for the first page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Values above the maximum are clamped, not rejected.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    RateLimitLimit:
      description: Request quota for the current window.
      schema:
        type: integer
    RateLimitRemaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: Unix epoch seconds when the window resets.
      schema:
        type: integer
    RetryAfter:
      description: Seconds to wait before retrying (rate limit or temporary degradation).
      schema:
        type: integer
...
  responses:
...
    TooManyRequests:
      description: Rate limit exceeded (edge, OTP throttle, or messaging rate limit).
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            rateLimited:
              $ref: '#/components/examples/ErrRateLimited'
    ServiceDegraded:
      description: A downstream dependency is degraded; a reduced result or retry hint is returned (search→DB,
        chat→click-to-call, config→last-good snapshot).
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            degraded:
              $ref: '#/components/examples/ErrDegraded'
  schemas:
    LocalizedText:
      type: object
      description: 'User-visible text across the four locales (DB PD-03). `uz_latn` is the canonical Uzbek;
        the

        others are optional depending on what was authored. Cross-script Uzbek matching is handled by

        the search engine, not by stored transliteration.

        '
      properties:
        uz_latn:
          type: string
        uz_cyrl:
          type: string
        ru:
          type: string
        en:
          type: string
      additionalProperties: false
      examples:
      - uz_latn: Kvartira
        ru: Квартира
        en: Apartment
    Money:
      type: object
      required:
      - amount
      - currency
      properties:
        amount:
          type: string
          description: Decimal string, 2 dp (avoids float error).
          examples:
          - '1500000.00'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          default: UZS
          examples:
          - UZS
      additionalProperties: false
    GeoPoint:
      type: object
      required:
      - latitude
      - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
      additionalProperties: false
...
    CursorPage:
      type: object
      description: Cursor pagination envelope (opaque forward cursor).
      required:
      - items
      - page
      properties:
        items:
          type: array
          items: {}
        page:
          type: object
          required:
          - limit
          properties:
            limit:
              type: integer
            nextCursor:
              type:
              - string
              - 'null'
              description: Pass as `cursor` to fetch the next page; null when exhausted.
            total:
              type:
              - integer
              - 'null'
              description: Present only where cheap to compute; may be null.
      additionalProperties: false
    Error:
      type: object
      description: 'Single Problem-style error envelope (RFC-9457 shaped). `code` is a stable, machine-readable

        member of a closed vocabulary and refines the HTTP status.

        '
      required:
      - type
      - title
      - status
      - code
      - traceId
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
          examples:
          - https://errors.activehome.uz/validation
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        code:
          type: string
          description: Stable application-level error code.
          enum:
          - VALIDATION_FAILED
          - BUSINESS_RULE_VIOLATION
          - AUTHENTICATION_REQUIRED
          - AUTHENTICATION_INVALID
          - SESSION_EXPIRED
          - PERMISSION_DENIED
          - WRONG_ACTING_PROFILE
          - RESOURCE_NOT_FOUND
          - CONFLICT
          - DUPLICATE_KEY
          - ILLEGAL_STATE_TRANSITION
          - QUOTA_EXCEEDED
          - IMAGE_COUNT_EXCEEDED
          - UNSUPPORTED_MEDIA_TYPE
          - IDEMPOTENCY_KEY_REUSED
          - RATE_LIMITED
          - OTP_THROTTLED
          - DEPENDENCY_DEGRADED
        detail:
          type: string
          description: Human-readable explanation for this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference for this specific occurrence.
        traceId:
          type: string
          description: Correlation id for support and logs.
        details:
          type: array
          description: Field-level errors (validation and business-rule cases).
          items:
            $ref: '#/components/schemas/ValidationError'
      additionalProperties: false
    ValidationError:
      type: object
      required:
      - path
      - rule
      - message
      properties:
        path:
          type: string
          description: JSON pointer / field path.
          examples:
          - /attributes/area_m2
        rule:
          type: string
          description: The whitelisted validator or schema rule that failed.
          examples:
          - numeric_range
        message:
          type: string
      additionalProperties: false
...
    SearchRequest:
      type: object
      description: Structured search query (also expressible via query parameters on GET /search).
      properties:
        q:
          type: string
          description: Free text; matched cross-script (Latin↔Cyrillic).
        categoryId:
          type: string
          format: uuid
        listingType:
          type: string
          enum:
          - ADVERTISEMENT
          - PRODUCT
          - SERVICE
        filters:
          type: object
          description: Facet filters keyed by facet-eligible field code (from the SearchConfiguration
            snapshot).
          additionalProperties: true
        priceMin:
          type: string
        priceMax:
          type: string
        verifiedOnly:
          type: boolean
        geo:
          type:
          - object
          - 'null'
          properties:
            center:
              $ref: '#/components/schemas/GeoPoint'
            radiusKm:
              type: number
              minimum: 0.1
              maximum: 200
          additionalProperties: false
        sort:
          type: string
          enum:
          - RELEVANCE
          - RECENCY
          - PRICE_ASC
          - PRICE_DESC
          default: RELEVANCE
      additionalProperties: false
    SearchResult:
      type: object
      description: Search response — matched listings, facet results, and promotion labelling.
      required:
      - items
      - facets
      - page
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
        facets:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        page:
          $ref: '#/components/schemas/CursorPage'
        degraded:
          type: boolean
          default: false
          description: True when served from the DB fallback (reduced result set).
      additionalProperties: false
    SearchHit:
      type: object
      required:
      - listingId
      - title
      properties:
        listingId:
          type: string
          format: uuid
        title:
          type: string
        categoryPath:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        location:
          $ref: '#/components/schemas/GeoPoint'
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
        verifiedBadge:
          type: boolean
        promoted:
          type:
          - object
          - 'null'
          description: Present and labelled when the hit is a capped promoted result (BRULE-10).
          properties:
            kind:
              type: string
              enum:
              - PREMIUM
              - FEATURED
              - TOP_PLACEMENT
          additionalProperties: false
        slug:
          type: string
      additionalProperties: false
    Facet:
      type: object
      required:
      - fieldCode
      - label
      - buckets
      properties:
        fieldCode:
          type: string
        label:
          $ref: '#/components/schemas/LocalizedText'
        buckets:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
            additionalProperties: false
      additionalProperties: false
    Suggestion:
      type: object
      required:
      - text
      properties:
        text:
          type: string
        type:
          type: string
          enum:
          - QUERY
          - CATEGORY
          - LISTING
        refId:
          type:
          - string
          - 'null'
          format: uuid
      additionalProperties: false
...
  examples:
...
    ErrRateLimited:
      summary: Rate limit exceeded
      value:
        type: https://errors.activehome.uz/rate-limit
        title: Too many requests
        status: 429
        code: RATE_LIMITED
        detail: Retry after the indicated interval.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a71
    ErrDegraded:
      summary: Dependency degraded
      value:
        type: https://errors.activehome.uz/degraded
        title: Service temporarily degraded
        status: 503
        code: DEPENDENCY_DEGRADED
        detail: Search is running in reduced (database fallback) mode.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a72
...
    SearchResultExample:
      summary: Faceted search response
      value:
        items:
        - listingId: 7b3f1c2a-0e4d-4a9b-8c1f-2d3e4f5a6b7c
          title: 3-room apartment in Yunusabad
          categoryPath: real-estate/residential/apartments
          price:
            amount: '1500000.00'
            currency: UZS
          location:
            latitude: 41.3667
            longitude: 69.287
          thumbnailUrl: https://cdn.activehome.uz/t/aa11bb22.webp
          verifiedBadge: true
          promoted:
            kind: FEATURED
          slug: 3-room-apartment-yunusabad-7b3f1c2a
        facets:
        - fieldCode: rooms
          label:
            uz_latn: Xonalar
            ru: Комнаты
            en: Rooms
          buckets:
          - value: '1'
            count: 120
          - value: '2'
            count: 240
          - value: '3'
            count: 180
        page:
          items: []
          page:
            limit: 20
            nextCursor: eyJvZmZzZXQiOjIwfQ
            total: 540
        degraded: false
...